refactor: make collections internal
