+++ .app/_config/collections/_notes.collection.js
